Support floatX=float32; use it in tests by default
fixes issue #28

+++ nips_madness/gradient_expressions/make_w_batch.py
@@ -2,7 +2,7 @@ import theano
 import theano.tensor as T
 import numpy as np
 
-sign = np.array([[1, -1], [1, -1]])
+sign = np.array([[1, -1], [1, -1]], dtype='int16')  # for floatX=float32
 dsign = sign.reshape([2, 2, 1, 1])
 
 def make_W_with_x(Z,J,D,S,N,X):
